created data_check.py
Added a levenes test + shapiro test that returns a boolean if they pass the tests. I'm a little unfamiliar with working with python so I couldn't test it with the example data, but I did test using my own and it works! Let me know if it doesn't work with your data and we can figure out a solution!